If there are no bugs ...
the database, dbmega2.db, is the identical for all analysis with a given data format.  All checks on analysis and on ReOrderLoci processing are done after the database is read in.  Any checks/cleaning before the database is dumped are on all loci.  The big change was splitting out or connecting pedigree loops where necessary.